+++ Izayoi.Data.DbCommandAdapter/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
 
+## [1.1.0] - 2024-08-22
+
+- Changes
+  - For the `DbCommandAdapter` constructor, the type of the argument `DbDataMapper` has been changed to `IDbDataMapper`.
+- Adds
+  - Added `IDbDataMapper` property to `DbCommandAdapter` and `IDbCommandAdapter`.
+
 ## [1.0.0] - 2024-08-16
 Initial release.